apps/data_chat_bot/backend/requirements.txt
# Data Chat Bot - Backend Dependencies
# ======================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# AI
openai==1.12.0

# CORS
python-multipart==0.0.6

# Environment
python-dotenv==1.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
